perf: don't always purge expired txs when checking for existence (#1488)
* perf: don't always purge expired txs when checking for existence

Before we check whether a transaction exists in the pool we remove
expired ones to avoid false positive - pretending that an expired
transaction exists in the pool.

However, if a transaction is not in the pool there is no need to purge
in order provide a definite answer.

+++ packages/core-transaction-pool-mem/lib/connection.js
@@ -359,6 +359,12 @@ class TransactionPool extends TransactionPoolInterface {
    * @return {Boolean}
    */
   transactionExists(transactionId) {
+    if (!this.mem.transactionExists(transactionId)) {
+      // If it does not exist then no need to purge expired transactions because
+      // we know it will not exist after purge too.
+      return false
+    }
+
     this.__purgeExpired()
 
     return this.mem.transactionExists(transactionId)
